requirements.txt
click==8.0.3
colour-science==0.4.1
cycler==0.11.0
Flask==2.0.2
Flask-Cors==3.0.10
fonttools==4.31.2
gunicorn==20.1.0
igraph==0.9.10
imageio==2.16.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.4.2
lsq-ellipse==2.2.1
MarkupSafe==2.0.1
matplotlib==3.5.1
networkx==2.8
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.7
python-dateutil==2.8.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
texttable==1.6.4
threadpoolctl==3.1.0
typing-extensions==4.1.1
Werkzeug==2.0.3
